Move MouseHelper into browser
